Add `draw_delay` option for all objects to delay spawn
Previously only "explosion" type supported delayed spawn and it was
randomized (up to 5 second delay). It was not possible to control it, it
was always on for these objects. This enables an additional parameter
which allows specifying exact delay for all object types and removes the
randomized spawn time delay for "explosion" type.

Closes: #1