Add `draw_delay` option for all objects to delay spawn (#5)
* Add `draw_delay` option for all objects to delay spawn

Previously only "explosion" type supported delayed spawn and it was
randomized (up to 5 second delay). It was not possible to control it, it
was always on for these objects. This enables an additional parameter
which allows specifying exact delay for all object types and removes the
randomized spawn time delay for "explosion" type.

Besides the changes from #1, this makes a small change to how arcs are drawn: previously the parts that were not drawn were black instead of being invisible, this PR fixes that too.

Closes: #1